requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
python-dotenv==0.19.2
requests==2.26.0
web3==5.31.1
py-solc-x==1.2.0
